+++ bindings/python/iota_sdk/utils.py
@@ -220,11 +220,11 @@ class Utils():
         })
 
     @staticmethod
-    def verify_semantic(
+    def verify_transaction_semantic(
             inputs: List[InputSigningData], transaction: TransactionPayload, time: int) -> bool:
         """Verifies the semantic of a transaction.
         """
-        return _call_method('verifySemantic', {
+        return _call_method('verifyTransactionSemantic', {
             'inputs': [i.as_dict() for i in inputs],
             'transaction': transaction.as_dict(),
             'time': time,