+++ bindings/python/CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- `Utils:verify_semantic()`;
+- `Utils:verify_transaction_semantic()`;
 
 ## 1.1.1 - 2023-10-31